requirements.txt
PyPDF2==3.0.1
fastapi==0.103.1
langchain==0.0.352
faiss-cpu==1.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
huggingface-hub==0.36.0
openai~=2.2.0
pandas~=2.0.3
uvicorn~=0.33.0
pydantic~=2.10.6
langchain-core~=0.1.23
torch~=2.4.1
tqdm~=4.67.1
transformers~=4.40.2
rouge-score~=0.1.2
nltk~=3.9.1
bert-score~=0.3.13
matplotlib~=3.7.5
icecream~=2.1.8